+++ src/day06.rs
@@ -1,12 +1,12 @@
-pub fn part1(input: String) -> u32 {
+pub fn part1(input: String) -> u64 {
     populate_laternfish(&input, 80)
 }
 
-pub fn part2(input: String) -> u32 {
+pub fn part2(input: String) -> u64 {
     populate_laternfish(&input, 256)
 }
 
-pub fn populate_laternfish(input: &String, days: u16) -> u32 {
+pub fn populate_laternfish(input: &String, days: u16) -> u64 {
     let mut population: Vec<u8> = input.lines().next().unwrap().split(',')
         .map(|c| c.parse::<u8>().unwrap()).collect();
 
@@ -26,8 +26,7 @@ pub fn populate_laternfish(input: &String, days: u16) -> u32 {
         println!("Day {:?}, {:?} total fish", day, fishes.iter().sum::<u64>());
     }
 
-    //the returned result will be incorrect due to cast but we can read it from output
-    fishes.iter().sum::<u64>() as u32
+    fishes.iter().sum::<u64>()
 }
 
 pub fn sub_modular(val: u8, sub: u8, base: u8) -> usize {
