+++ src/day07.rs
@@ -1,12 +1,12 @@
-pub fn part1(input: String) -> u32 {
+pub fn part1(input: String) -> u64 {
     find_optimal_position(&input, false)
 }
 
-pub fn part2(input: String) -> u32 {
+pub fn part2(input: String) -> u64 {
     find_optimal_position(&input, true)
 }
 
-pub fn find_optimal_position(input: &String, increasing: bool) -> u32 {
+pub fn find_optimal_position(input: &String, increasing: bool) -> u64 {
     let positions: Vec<i16> = input.split(',')
         .map(|c| c.parse::<i16>().unwrap()).collect();
     let mut fuel_sums: Vec<u32> = Vec::new();
@@ -22,5 +22,5 @@ pub fn find_optimal_position(input: &String, increasing: bool) -> u32 {
             });
         fuel_sums.push(sum);
     }
-    *fuel_sums.iter().min().unwrap()
+    *fuel_sums.iter().min().unwrap() as u64
 }
